Fix NaN values in FoldKe and PathAnalysis

+++ python/GetData.py
@@ -17,7 +17,7 @@ def GetData(Ui, Node, truss, angles):
     
     for bel in range(truss['Bars'].shape[0]):
         eDof = np.array([np.arange(-2, 1) + truss['Bars'][bel, 0] * 3,
-                        np.arange(-2, 1) + truss['Bars'][bel, 1] * 3]).T
+                        np.arange(-2, 1) + truss['Bars'][bel, 1] * 3], dtype="int").T
         Exbar[bel], Rbe, Kbe= BarKe(Ui[eDof.ravel()], csr_matrix(truss['B'])[bel, eDof.ravel()], truss['L'][bel], truss['CM'], truss['A'][bel])
     
     for d_el in range(angles['bend'].shape[0]):
